# packages/server/docs/swagger.yaml
openapi: 3.0.0
info:
  title: '@resala/server'
  version: 1.0.0
  description: Resala server
  license:
    name: ISC
  contact:
    name: Hossam Hamdy
paths:
  /api/v1/users/{userId}:
    get:
      operationId: GetUser
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserResponse'
      tags:
        - User
      security:
        - jwt_auth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
    delete:
      operationId: DeleteUser
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteUserResponse'
      tags:
        - User
      security:
        - jwt_auth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
    put:
      operationId: UpdateUser
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateUserResponse'
      tags:
        - User
      security:
        - jwt_auth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                role:
                  type: string
                  enum:
                    - ADMIN
                    - CUSTOMER
                    - MODERATOR
                lastName:
                  type: string
                firstName:
                  type: string
                phone:
                  type: string
                isVerified:
                  type: boolean
              type: object
  /api/v1/users:
    get:
      operationId: ListUsers
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUsersResponse'
      tags:
        - User
      security:
        - jwt_auth: []
      parameters:
        - in: query
          name: page
          required: false
          schema:
            default: 1
            format: int32
            type: integer
        - in: query
          name: limit
          required: false
          schema:
            default: 10
            format: int32
            type: integer
        - in: query
          name: search
          required: false
          schema:
            type: string
...
components:
...
  schemas:
    DefaultResponseBody:
      properties:
        message:
          type: string
        success:
          type: boolean
      required:
        - success
      type: object
    RoleType:
      type: string
      enum:
        - ADMIN
        - CUSTOMER
        - MODERATOR
      nullable: false
    User:
      properties:
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
          nullable: true
        role:
          $ref: '#/components/schemas/RoleType'
        lastName:
          type: string
        firstName:
          type: string
        isPhoneVerified:
          type: boolean
        phone:
          type: string
          nullable: true
        isEmailVerified:
          type: boolean
        email:
          type: string
        id:
          type: string
      required:
        - updatedAt
        - createdAt
        - lastLogin
        - role
        - lastName
        - firstName
        - isPhoneVerified
        - phone
        - isEmailVerified
        - email
        - id
      type: object
    GetUserResponse:
      allOf:
        - $ref: '#/components/schemas/DefaultResponseBody'
        - properties:
            data:
              $ref: '#/components/schemas/User'
          required:
            - data
          type: object
    Pagination:
      properties:
        total:
          type: integer
          format: int32
        limit:
          type: integer
          format: int32
        page:
          type: integer
          format: int32
      required:
        - total
        - limit
        - page
      type: object
    ListUsersResponse:
      allOf:
        - $ref: '#/components/schemas/DefaultResponseBody'
        - properties:
            data:
              properties:
                users:
                  items:
                    $ref: '#/components/schemas/User'
                  type: array
                pagination:
                  $ref: '#/components/schemas/Pagination'
              required:
                - users
                - pagination
              type: object
          required:
            - data
          type: object
    DeleteUserResponse:
      $ref: '#/components/schemas/DefaultResponseBody'
    UpdateUserResponse:
      allOf:
        - $ref: '#/components/schemas/DefaultResponseBody'
        - properties:
            data:
              $ref: '#/components/schemas/User'
          required:
            - data
          type: object
...
  securitySchemes:
    jwt_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://api.resala.live
    description: Development server